Floodlight topology information retrieval to Topology Bridge
This commit contains the initial code that enables OSCARS's
Topology Bridge module to retrieve topology from Floodlight.
The code can be easily extended to incorporate information
from multiple SDN controllers. At the core are the
ISDNTopologyService interface, which defines what the
Topology Service should produce when the module requests
the topology and the BaseSDNTopologyService class, which
implements the singleton logic.

TODO:
Finish the code that exports the topology using XML
Fix maven dependencies for Jersey and Jackson libs
Add Jersey and Jackson jars to the repository