Implements the concept of Entry Hop and Exit Hop
There are two special types of SDNHops: entry hops and exit hops.
This commit implements the logic for these two special cases.

Suppose that we're given a circuit reservation like:

A.port1,A.port2,...,B.port1,B.port2

In this case, all the hops are "complete", i.e. there is no hop with only one
port specified in the reservation. This means that all traffic coming from
A.port1 should be forwarded to B.port2 no matter what, regardless of flow
distinction.

Now supposed that we're given a reservation like:
A.port2,D.port1,D.port2,...,C.port1,C.port2,B.port1

What traffic should be forwarded through the circuit in this case? With the
SDN PSS implementation, the hop A is said to be an Entry Hop to the circuit,
and B is said to be an Exit Hop. The traffic that will be forwarded through
the circuit will depend on the existence of an OFMatch specified by the user.
If the user specifies an OFMatch, then that will be used to choose what
traffic will flow through the circuit. However, if no OFMatch is specified,
the SDN PSS leaves this decision to the switch and don't install any rules on
A or B.